requirements.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
email_validator==1.1.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==0.14.1
WTForms==2.2.1
